requirements.txt
boto3==1.16.56
botocore==1.19.56
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
netaddr==0.8.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.2
